% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classific.R
\name{classific}
\alias{classific}
\title{Classificação}
\usage{
classific(bd, rodada)
}
\arguments{
\item{bd}{df salvo no pacote}

\item{rodada}{rodada para gerar a classificacao.}
}
\value{
a classificação no início da rodada.
}
\description{
Essa função gera a classificação da rodada, fiz isso para não importar mais
uma base de dados, podemos chamar como uma funçãoo independente.
}
