% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/otimizado.R
\name{prevCBA}
\alias{prevCBA}
\title{prevCBA}
\usage{
prevCBA(data, rodada, meanCBPfm, meanCBPsm, meanCBPfv, meanCBPsv, nsim, CopaBR, Lib)
}
\arguments{
\item{data}{banco de dados}

\item{rodada}{rodada}

\item{meanCBPfm}{meanCBPfm}

\item{meanCBPsm}{meanCBPsm}

\item{meanCBPfv}{meanCBPfv}

\item{meanCBPsv}{meanCBPsv}

\item{nsim}{numero de simulacoes}
}
\description{
Faz toda a previsão
}
